requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.0.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbformat==5.0.6
numpy==1.18.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
QtPy==1.9.0
requests==2.23.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.1.0
